tests for settings and featflags
