+++ tests/openbb_terminal/stocks/insider/cassettes/test_openinsider_view/test_print_insider_filter.yaml
@@ -166,7 +166,7 @@ interactions:
       Content-Type:
       - text/html; charset=UTF-8
       Date:
-      - Sat, 21 May 2022 14:38:18 GMT
+      - Wed, 25 May 2022 00:15:15 GMT
       Keep-Alive:
       - timeout=5, max=10
       Server:
