+++ tests/openbb_terminal/test_settings_controller.py
@@ -21,68 +21,8 @@ def test_print_help(controller):
     controller.print_help()
 
 
-def test_call_logcollection(controller):
-    controller.call_logcollection(None)
-
-
-def test_call_tab(controller):
-    controller.call_tab(None)
-
-
-def test_call_cls(controller):
-    controller.call_cls(None)
-
-
-def test_call_color(controller):
-    controller.call_color(None)
-
-
-def test_call_promptkit(controller):
-    controller.call_promptkit(None)
-
-
-def test_call_predict(controller):
-    controller.call_predict(None)
-
-
-def test_call_thoughts(controller):
-    controller.call_thoughts(None)
-
-
-def test_call_reporthtml(controller):
-    controller.call_reporthtml(None)
-
-
-def test_call_exithelp(controller):
-    controller.call_exithelp(None)
-
-
-def test_call_rcontext(controller):
-    controller.call_rcontext(None)
-
-
 def test_call_dt(controller):
     controller.call_dt(None)
-
-
-def test_call_rich(controller):
-    controller.call_rich(None)
-
-
-def test_call_richpanel(controller):
-    controller.call_richpanel(None)
-
-
-def test_call_ion(controller):
-    controller.call_ion(None)
-
-
-def test_call_watermark(controller):
-    controller.call_watermark(None)
-
-
-def test_call_cmdloc(controller):
-    controller.call_cmdloc(None)
 
 
 def test_call_autoscaling(controller):
